feat: :sparkles: Update user model, create property and property rating models and seed data
This commit introduces updates to the user entity by adding createdAt and updatedAt fields, and creates new entities for property and propertyRating. Additionally, it includes seed data for the newly created entities to facilitate initial testing and development environments.

Important changes include:  Enhancement of the user entity for better data tracking. Introduction of property and propertyRating entities to extend the application's functionality. Provision of seed data for new entities to support development and testing processes.