feat: :seedling: Clear Algolia index before seeding data
This commit introduces the clearing of the Algolia index before seeding new data to ensure the seeded data is always fresh and relevant. This change improves the reliability and accuracy of the seeded data in the development environment.

Important changes: - Algolia index is cleared before seeding data, ensuring only the   most current data is present.